feat(Uri): Add require_network() function for network validation
